requirements.txt
airbyte==0.11.0
pandas==2.1.4
boto3==1.40.17
python-dotenv==1.0.0
pyarrow==21.0.0